+++ ui/src/routes/download/index.svelte
@@ -1,0 +1,7 @@
+<script context="module">
+	import {basename} from 'app/utils';
+
+	export async function preload() {
+		return this.redirect(302, `/data/${basename}.zip`);
+	}
+</script>